requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pytest
httpx
email-validator
tldextract
beautifulsoup4
openai
requests